Clarify interpreter executor metrics semantics (#66)
Add `observe_value` for non-latency CPU executor measurements and keep
`observe_latency_ms` reserved for latency observations.

Move duplicated executor test helpers into shared test configuration,
make shutdown idempotency explicit, and simplify the private worker runtime
parser docstring.
